+++ packages/gatsby/src/query/index.js
@@ -97,6 +97,9 @@ const popNodeQueries = state => {
 
     return dirtyIds
   }, new Set())
+
+  boundActionCreators.deleteComponentsDependencies([...uniqDirties])
+
   queuedDirtyActions = []
   return uniqDirties
 }