+++ packages/gatsby/src/utils/develop-proxy.ts
@@ -1,7 +1,9 @@
 import { createServer } from "http"
 import httpProxy from "http-proxy"
+import path from "path"
 import fs from "fs-extra"
 import { getServices } from "gatsby-core-utils/dist/service-lock"
+import st from "st"
 import restartingScreen from "./restarting-screen"
 
 interface IProxyControls {
@@ -11,6 +13,17 @@ interface IProxyControls {
 }
 
 const noop = (): void => {}
+
+const adminFolder = path.join(
+  path.dirname(require.resolve(`gatsby-admin`)),
+  `public`
+)
+
+const serveAdmin = st({
+  path: adminFolder,
+  url: `/___admin`,
+  index: `index.html`,
+})
 
 export const startDevelopProxy = (input: {
   proxyPort: number
@@ -32,6 +45,13 @@ export const startDevelopProxy = (input: {
   proxy.on(`error`, noop)
 
   const server = createServer((req, res) => {
+    if (process.env.GATSBY_EXPERIMENTAL_ENABLE_ADMIN) {
+      const wasAdminRequest = serveAdmin(req, res)
+      if (wasAdminRequest) {
+        return
+      }
+    }
+
     // Add a route at localhost:8000/___services for service discovery
     if (req.url === `/___services`) {
       getServices(input.programPath).then(services => {