+++ packages/gatsby-remark-code-repls/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-code-repls@3.2.2...gatsby-remark-code-repls@3.2.3) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-remark-code-repls
+
 ## [3.2.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-code-repls@3.2.1...gatsby-remark-code-repls@3.2.2) (2020-05-20)
 
 **Note:** Version bump only for package gatsby-remark-code-repls