+++ packages/gatsby-source-graphql/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@2.5.2...gatsby-source-graphql@2.5.3) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
 ## [2.5.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@2.5.1...gatsby-source-graphql@2.5.2) (2020-05-20)
 
 **Note:** Version bump only for package gatsby-source-graphql